KS-11-2 fix issues
